+++ selfdrive/car/hyundai/carcontroller.py
@@ -5,7 +5,7 @@ from common.realtime import DT_CTRL
 from opendbc.can.packer import CANPacker
 from selfdrive.car import apply_std_steer_torque_limits
 from selfdrive.car.hyundai import hda2can, hyundaican
-from selfdrive.car.hyundai.values import Buttons, CarControllerParams, HDA2_CAR, CAR
+from selfdrive.car.hyundai.values import Buttons, CarControllerParams, HDA2_CAR, CAR, CAN_FD_CAR
 
 VisualAlert = car.CarControl.HUDControl.VisualAlert
 LongCtrlState = car.CarControl.Actuators.LongControlState
@@ -54,7 +54,7 @@ class CarController:
 
     # These cars have significantly more torque than most HKG.  Limit to 70% of max.
     steer = actuators.steer
-    if self.CP.carFingerprint in (CAR.KONA, CAR.KONA_EV, CAR.KONA_HEV):
+    if self.car_fingerprint in (CAR.KONA, CAR.KONA_EV, CAR.KONA_HEV):
       steer = clip(steer, -0.7, 0.7)
     new_steer = int(round(steer * self.params.STEER_MAX))
     apply_steer = apply_std_steer_torque_limits(new_steer, self.apply_steer_last, CS.out.steeringTorque, self.params)
@@ -69,23 +69,23 @@ class CarController:
 
     can_sends = []
 
-    if self.CP.carFingerprint in HDA2_CAR:
+    if self.car_fingerprint in HDA2_CAR or CAN_FD_CAR:
       # steering control
-      can_sends.append(hda2can.create_lkas(self.packer, CC.enabled, CC.latActive, apply_steer))
+      can_sends.append(hda2can.create_lkas(self.packer, self.car_fingerprint, CC.enabled, CC.latActive, apply_steer))
 
-      if self.frame % 5 == 0:
+      if self.frame % 5 == 0 and self.car_fingerprint in HDA2_CAR:
         can_sends.append(hda2can.create_cam_0x2a4(self.packer, CS.cam_0x2a4))
 
       # cruise cancel
       if (self.frame - self.last_button_frame) * DT_CTRL > 0.25:
         if CC.cruiseControl.cancel:
           for _ in range(20):
-            can_sends.append(hda2can.create_buttons(self.packer, CS.buttons_counter+1, Buttons.CANCEL))
+            can_sends.append(hda2can.create_buttons(self.packer, self.car_fingerprint, CS.cruise_buttons_copy, CS.buttons_counter+1, Buttons.CANCEL))
           self.last_button_frame = self.frame
 
         # cruise standstill resume
         elif CC.cruiseControl.resume:
-          can_sends.append(hda2can.create_buttons(self.packer, CS.buttons_counter+1, Buttons.RES_ACCEL))
+          can_sends.append(hda2can.create_buttons(self.packer, self.car_fingerprint, CS.cruise_buttons_copy, CS.buttons_counter+1, Buttons.RES_ACCEL))
           self.last_button_frame = self.frame
     else: